requirements.txt
Django>=3.2,<5.0
djangorestframework>=3.13.1,<3.15.1
django-model-utils==4.2.0
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.26.0,<0.27
